// de-care/libs/de-care-use-cases/account/state-my-account-subscriptions/src/lib/state/interface.ts
export type SubscriptionCardComponent =
    | 'ACTIVE_SUBSCRIPTION'
    | 'INACTIVE_SUBSCRIPTION'
    | 'INACTIVE_SUBSCRIPTION_DUE_TO_SEASONAL_SUSPENSION'
    | 'INACTIVE_STREAMING_SUBSCRIPTION_DUE_TO_SEASONAL_SUSPENSION'
    | 'AVIATION_SUBSCRIPTION'
    | 'MARINE_SUBSCRIPTION'
    | 'ADD_SUBSCRIPTION'
    | 'NEW_SUBSCRIPTION'
    | 'STREAMING_SUBSCRIPTION'
    | 'INACTIVE_STREAMING_SUBSCRIPTION'
    | 'DONT_SEE_YOUR_SUBSCRIPTION';